+++ app/app.py
@@ -1,21 +1,26 @@
-from collections import defaultdict
-import json
-from functools import wraps
-from typing import Any
 from flask import Flask, request, Response, abort, render_template, send_file, jsonify
-from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
-from datetime import datetime
-from mattermostdriver import Driver
-import requests
-import config
-import random
+from flask_sqlalchemy import SQLAlchemy
+from functools import wraps
 import hashlib
 import hmac
+import json
+from mattermostdriver import Driver
 import re
-import sys
+import requests
 import time
-import traceback
+
+import config
+from routes import (
+    cammie_blueprint,
+    door_access_blueprint,
+    door_control_blueprint,
+    doorkeeper_blueprint,
+    fingerprint_blueprint,
+    quote_blueprint,
+    resto_blueprint,
+    spaceapi_blueprint,
+)
 
 
 app = Flask(__name__)
@@ -25,6 +30,15 @@ app.config["SQLALCHEMY_DATABASE_URI"] = config.DATABASE_URL
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
+
+app.register_blueprint(cammie_blueprint)
+app.register_blueprint(door_access_blueprint)
+app.register_blueprint(door_control_blueprint)
+app.register_blueprint(doorkeeper_blueprint)
+app.register_blueprint(fingerprint_blueprint)
+app.register_blueprint(quote_blueprint)
+app.register_blueprint(resto_blueprint)
+app.register_blueprint(spaceapi_blueprint)
 
 response_setting = "in_channel"
 
@@ -123,77 +137,6 @@ def get_actual_username(username):
     return username.lstrip("@")
 
 
-@app.route("/authorize", methods=["POST"])
-@requires_token("authorize")
-@requires_admin
-def authorize(admin_user):
-    """Slash-command to authorize a new user or modify an existing user"""
-    tokens = request.values.get("text").strip().split()
-    if not tokens:
-        # list authorized user
-        response = "\n".join(
-            f'{"**" if u.admin else ""}{u.username}{" ADMIN**" if u.admin else ""}'
-            for u in models.User.query.filter_by(authorized=True).order_by(
-                models.User.username
-            )
-        )
-        return mattermost_response(response, ephemeral=True)
-    if len(tokens) > 2:
-        return mattermost_response(
-            "To authorize a user: /authorize username [admin]\nTo list authorized users: /authorize",
-            ephemeral=True,
-        )
-    to_authorize_username = get_actual_username(tokens[0])
-    to_authorize_id = get_mattermost_id(to_authorize_username)
-    if to_authorize_id is None:
-        return mattermost_response(
-            "User '{}' does not seem to exist in Mattermost".format(
-                to_authorize_username
-            ),
-            ephemeral=True,
-        )
-    as_admin = len(tokens) == 2 and tokens[1] == "admin"
-    user = models.User.query.filter_by(mattermost_id=to_authorize_id).first()
-    if not user:
-        user = models.User(to_authorize_username)
-        user.mattermost_id = to_authorize_id
-    user.authorized = True
-    user.admin = as_admin or user.admin
-    db.session.add(user)
-    db.session.commit()
-    if user.admin:
-        return mattermost_response("'{}' is now an admin".format(to_authorize_username))
-    else:
-        return mattermost_response(
-            "'{}' is now a regular user".format(to_authorize_username)
-        )
-
-
-@app.route("/revoke", methods=["POST"])
-@requires_token("revoke")
-@requires_admin
-def revoke(admin_username):
-    """Slash-command to revoke a user"""
-    tokens = request.values.get("text").strip().split()
-    to_revoke_username = get_actual_username(tokens[0])
-    to_revoke_id = get_mattermost_id(to_revoke_username)
-    if to_revoke_id is None:
-        return mattermost_response(
-            "Could not find '{}' in Mattermost".format(to_revoke_username)
-        )
-    user = models.User.query.filter_by(mattermost_id=to_revoke_id).first()
-    if user is None:
-        return mattermost_response(
-            "Could not find '{}' in our database".format(to_revoke_username)
-        )
-    if user.admin:
-        return mattermost_response("Can't revoke admin user")
-    user.authorized = False
-    db.session.add(user)
-    db.session.commit()
-    return mattermost_response("'{}' revoked".format(to_revoke_username))
-
-
 def lockbot_request(command):
     # TODO: fix this properly with a mutex
     # TODO: cache status requests
@@ -206,405 +149,6 @@ def lockbot_request(command):
     )
     r = requests.post(config.lockbot_url, payload, headers={"HMAC": calculated_hmac})
     return DOOR_STATUS[r.text]
-
-
-@app.route("/door", methods=["POST"])
-@requires_token("door")
-@requires_regular
-def door(user):
-    tokens = request.values.get("text").strip().split()
-    command = tokens[0].lower()
-    if command == "getkey":
-        user.generate_key()
-        db.session.add(user)
-        db.session.commit()
-        return mattermost_response(
-            f"WARNING: door should only be operated when you are physically at the door. Your key is {user.doorkey}, the URLs you can POST to are https://mattermore.zeus.gent/api/door/{user.doorkey}/open and https://mattermore.zeus.gent/api/door/{user.doorkey}/lock",
-            ephemeral=True,
-        )
-    if command == "close":
-        command = "lock"
-    if command not in ("open", "lock", "status"):
-        return mattermost_response(
-            "Only [open|lock|status|getkey] subcommands supported", ephemeral=True
-        )
-    translated_state_before_command = lockbot_request(command)
-    if command != "status":
-        mattermost_doorkeeper_message(
-            f"door was {translated_state_before_command}, {user.username} tried to {command} door"
-        )
-    return mattermost_response(translated_state_before_command, ephemeral=True)
-
-
-@app.route("/api/door/<doorkey>/<command>", methods=["POST"])
-def doorapi(doorkey, command):
-    if not doorkey:
-        return abort(401)
-    user = models.User.query.filter_by(doorkey=doorkey, authorized=True).first()
-    if user is None:
-        return abort(401)
-    if command not in ("open", "lock", "status"):
-        return abort(400, "Command not in (open,lock,status)")
-    translated_state_before_command = lockbot_request(command)
-    if command != "status":
-        mattermost_doorkeeper_message(
-            f"door was {translated_state_before_command}, {user.username} tried to {command} door via the API"
-        )
-    return jsonify({"status": "ok", "before": translated_state_before_command})
-
-
-@app.route("/spaceapi.json")
-def spaceapi():
-    door_status = lockbot_request("status")
-    # Avoid XML parsing
-    status = door_status == "open"
-    response = jsonify(
-        {
-            "api": "0.13",
-            "space": "Zeus WPI",
-            "logo": "https://zinc.zeus.gent",
-            "url": "https://zeus.ugent.be",
-            "location": {
-                "address": "Zeuskelder, gebouw S9, Krijgslaan 281, Ghent, Belgium",
-                "lon": 3.7102741,
-                "lat": 51.0231119,
-            },
-            "contact": {"email": "bestuur@zeus.ugent.be", "twitter": "@ZeusWPI"},
-            "issue_report_channels": ["email"],
-            "state": {
-                "icon": {
-                    "open": "https://zinc.zeus.gent/zeus",
-                    "closed": "https://zinc.zeus.gent/black",
-                },
-                "open": status,
-            },
-            "projects": ["https://github.com/zeuswpi", "https://git.zeus.gent"],
-        }
-    )
-    response.headers.add("Access-Control-Allow-Origin", "*")
-    return response
-
-
-@app.route("/doorkeeper", methods=["POST"])
-def doorkeeper():
-    raw_data = request.get_data()
-    hmac_header = bytearray.fromhex(request.headers.get("HMAC"))
-    calculated_hash = hmac.new(
-        config.up_key.encode("utf8"), raw_data, hashlib.sha256
-    ).digest()
-    if hmac_header != calculated_hash:
-        print(f"WRONG: {hmac_header} != {calculated_hash}", file=sys.stderr)
-        return abort(401)
-
-    data_dict = {
-        l.split("=")[0]: l.split("=")[1] for l in raw_data.decode("utf8").split("&")
-    }
-    cmd = data_dict["cmd"]
-    reason = data_dict["why"]
-    value = data_dict["val"]
-    try:
-        requests.post(
-            config.kelderapi_doorkeeper_url,
-            json=data_dict,
-            headers={"Token": config.kelderapi_doorkeeper_key},
-            timeout=3,
-        )
-    except requests.exceptions.RequestException as e:
-        mattermost_doorkeeper_message(
-            f"Posting `{data_dict}` to kelderapi failed\n```\n{e.__class__.__name__}: {e}\n```",
-            webhook=config.debug_webhook,
-        )
-    except:
-        mattermost_doorkeeper_message(
-            f"Posting `{data_dict}` to kelderapi failed\n```\n{traceback.format_exc()}\n```",
-            webhook=config.debug_webhook,
-        )
-    if reason == "mattermore":
-        if cmd == "status":
-            return ""
-        msg = f'"{cmd}" command from Mattermost handled'
-    elif reason == "boot":
-        msg = "lockbot booted"
-    elif reason == "panic":
-        msg = f"@sysadmin: the door panicked with reason {cmd}"
-    elif reason == "state":
-        msg = f"The door is now {DOOR_STATUS[value]}"
-    elif reason == "chal":
-        return ""
-    elif reason == "delaybutton":
-        msg = "Delayed door close button was pressed"
-    else:
-        msg = f"Unhandled message type: {cmd},{reason},{value}"
-    mattermost_doorkeeper_message(msg, webhook=config.debug_webhook)
-    return "OK"
-
-
-@app.route("/cammiechat", methods=["POST"])
-@requires_token("cammiechat")
-@requires_regular
-def cammiechat(user):
-    headers = {"X-Username": user.username}
-    requests.post(
-        "https://kelder.zeus.ugent.be/messages/",
-        data=request.values.get("text").strip(),
-        headers=headers,
-        timeout=5,
-    )
-    return mattermost_response("Message sent", ephemeral=True)
-
-
-@app.route("/addquote", methods=["POST"])
-@requires_token("quote")
-def add_quote():
-    user = request.values["user_name"]
-    channel = request.values["channel_name"]
-    quote_text = request.values["text"]
-    quote = models.Quote(user, quote_text, channel)
-    db.session.add(quote)
-    db.session.commit()
-    return mattermost_response('{} added the quote "{}"'.format(user, quote_text))
-
-
-@app.route("/quote", methods=["POST"])
-def random_quote():
-    text_contains = request.values["text"]
-    matches = models.Quote.query.filter(
-        models.Quote.quote.contains(text_contains)
-    ).all()
-    if matches:
-        selected_quote = random.choice(matches)
-        response = selected_quote.quote
-        return mattermost_response(response)
-    return mattermost_response(
-        'No quotes found matching "{}"'.format(text_contains), ephemeral=True
-    )
-
-
-def fingerprint_request(command: str, data=None):
-    timestamp = int(time.time() * 1000)
-    payload = f"{timestamp};{command};{data};" if data else f"{timestamp};{command};"
-    calculated_hmac = (
-        hmac.new(config.down_key.encode("utf8"), payload.encode("utf8"), hashlib.sha256)
-        .hexdigest()
-        .upper()
-    )
-    return requests.post(
-        config.fingerprint_url, payload, headers={"HMAC": calculated_hmac}
-    )
-
-
-def get_free_fp_ids():
-    """Get a set of all available fingerprint IDs"""
-
-    res = fingerprint_request("list")
-
-    id_list = list(res.text)
-    used_ids = set()
-    for i, val in enumerate(id_list):
-        if val == "1":
-            used_ids.add(i)
-
-    all_ids = set(range(1, 201))
-    return all_ids.difference(used_ids)
-
-
-def pretty_user_fingerprints(d):
-    """Pretty print a dict of users and their fingerprints"""
-
-    repr = ""
-
-    for k, v in d.items():
-        repr += f"{k}\n"
-        for f in v:
-            repr += f"\t{f}\n"
-
-    return repr
-
-
-def send_fingerprint_delete(ids: list[int]):
-    """Mass delete a list of fingerprints"""
-
-    for id_ in ids:
-        fingerprint_request("delete", id_)
-
-
-@app.route("/fingerprint", methods=["POST"])
-@requires_token("fingerprint")
-@requires_regular
-def fingerprint(user):
-    tokens = request.values.get("text").strip().split()
-    try:
-        command = tokens[0].lower()
-    except IndexError:
-        return mattermost_response(
-            "Only [enroll|delete|list] subcommands supported", ephemeral=True
-        )
-
-    user_id = (
-        models.User.query.filter(models.User.username == user.username)
-        .from_self(models.User.id)
-        .scalar()
-    )
-
-    if command == "enroll":
-        try:
-            fp_note = tokens[1].lower()
-        except IndexError:
-            return mattermost_response(
-                "Missing or invalid fingerprint note, syntax: /fingerprint enroll \{note\}\n\{note\} must not contain spaces",
-                ephemeral=True,
-            )
-
-        # Ensure that no pending fingerprints remain (this only rarely do something)
-        deleted_ids = models.Fingerprint.clear_inactive(db)
-        send_fingerprint_delete(deleted_ids)
-
-        ids = get_free_fp_ids()
-        if len(ids) == 0:
-            return mattermost_response(
-                "Cannot enroll fingerprint, no free slots left", ephemeral=True
-            )
-
-        fp_id = min(ids)
-        fingerprint_request("enroll", fp_id)
-
-        models.Fingerprint.create(db, fp_id, user_id, fp_note, datetime.now())
-
-        print(
-            f"created inactive fingerprint {fp_id} for user {user.username} with note {fp_note}"
-        )
-        return mattermost_response(
-            f"Started enrolling fingerprint #{fp_id} for user '{user.username}'",
-            ephemeral=True,
-        )
-
-    if command == "delete":
-        try:
-            fp_note = tokens[1]
-        except IndexError:
-            return mattermost_response(
-                "Missing or invalid fingerprint note, syntax: /fingerprint delete \{note\}\n\{note\} must not contain spaces",
-                ephemeral=True,
-            )
-
-        fingerprint = models.Fingerprint.find(db, user_id, fp_note)
-
-        if fingerprint is None:
-            return mattermost_response(
-                f"No fingerprint with note '{fp_note}' found for user '{user.username}'",
-                ephemeral=True,
-            )
-
-        fingerprint_request("delete", fingerprint.id)
-
-        print(f"sent command delete fingerprint {fp_note} for {user.username}")
-        return mattermost_response(
-            f"Deleted fingerprint '{fp_note}' for user '{user.username}'",
-            ephemeral=True,
-        )
-
-    if command == "list":
-        data = (
-            db.session.query(models.User, models.Fingerprint)
-            .join(models.Fingerprint)
-            .all()
-        )
-        user_fingerprints = defaultdict(list)
-        for datum in data:
-            user_fingerprints[datum[0].username].append(datum[1].note)
-
-        if not user.admin:
-            if len(user_fingerprints[user.username]) != 0:
-                msg = user_fingerprints[user.username]
-            else:
-                msg = "No fingerprints found"
-
-            return mattermost_response(msg, ephemeral=True)
-
-        if len(user_fingerprints) == 0:
-            return mattermost_response("No fingerprints found", ephemeral=True)
-
-        return mattermost_response(
-            pretty_user_fingerprints(user_fingerprints), ephemeral=True
-        )
-
-    if command not in ("enroll", "delete", "list"):
-        return mattermost_response(
-            "Only [enroll|delete|list] subcommands supported", ephemeral=True
-        )
-
-    return mattermost_response("")
-
-
-@app.route("/fingerprint_cb", methods=["POST"])
-def fingerprint_cb():
-    msg, val = request.data.decode("utf-8").split("\n")
-
-    if msg == "enrolled":
-        fingerprint = models.Fingerprint.find_by_id(db, int(val))
-        fingerprint.active = True
-        fingerprint.save(db)
-        print(f"[ACK] activated fingerprint {fingerprint.id} for {fingerprint.user_id}")
-        mattermost_doorkeeper_message(
-            f"Activated fingerprint {fingerprint.note} for user {fingerprint.user.username}",
-            webhook=config.debug_webhook,
-        )
-    elif msg == "detected":
-        fingerprint = models.Fingerprint.find_active_by_id(db, int(val))
-        print(f"detected fingerprint {fingerprint.id}")
-        user = fingerprint.user
-
-        translated_state_before_command = lockbot_request("open")
-        mattermost_doorkeeper_message(
-            f"Detected fingerprint #{fingerprint.id} (user '{user.username}'",
-            webhook=config.debug_webhook,
-        )
-        mattermost_doorkeeper_message(
-            f"door was {translated_state_before_command}, {user.username} tried to open the door with the fingerprint sensor"
-        )
-    elif msg == "deleted":
-        fingerprint = models.Fingerprint.find_by_id(db, int(val))
-        if fingerprint is None:
-            return Response("", status=200)
-
-        note = fingerprint.note
-        user = fingerprint.user
-        user_id = fingerprint.user_id
-
-        fingerprint.delete_(db)
-        print(f"[ACK] deleted fingerprint '{note}' for '{user_id}'")
-        mattermost_doorkeeper_message(
-            f"Deleted fingerprint '{fingerprint.note}' for user '{user.username}'",
-            webhook=config.debug_webhook,
-        )
-
-    elif msg == "missing_hmac":
-        mattermost_doorkeeper_message(
-            "@sysadmin Fingerprint sensor received message without HMAC signature",
-            webhook=config.debug_webhook,
-        )
-    elif msg == "too_long":
-        mattermost_doorkeeper_message(
-            "@sysadmin Fingerprint sensor received message longer than 128 bytes",
-            webhook=config.debug_webhook,
-        )
-    elif msg == "invalid_hmac":
-        mattermost_doorkeeper_message(
-            "@sysadmin Fingerprint sensor received message with invalid HMAC signature",
-            webhook=config.debug_webhook,
-        )
-    elif msg == "replay":
-        mattermost_doorkeeper_message(
-            "@sysadmin Fingerprint sensor received message with incorrect timestamp (possible replay attack)",
-            webhook=config.debug_webhook,
-        )
-    else:
-        mattermost_doorkeeper_message(
-            "@sysadmin Received invalid fingerprint callback message",
-            webhook=config.debug_webhook,
-        )
-
-    return Response("", status=200)
 
 
 @app.route("/robots.txt", methods=["GET"])
@@ -647,92 +191,3 @@ def json_quotes():
     )
     response.headers.add("Access-Control-Allow-Origin", "*")
     return response
-
-
-RESTO_TEMPLATE = """
-# Restomenu
-
-## Woater me e smaksje
-{soup_table}
-
-## Vleesch
-{meat_table}
-
-## Visch
-{fish_table}
-
-## Vegetoarisch
-{vegi_table}
-{uncategorized}
-## Groensels
-{vegetable_table}
-
-"""
-
-UNCATEGORIZED_TEMPLATE = """
-## Nog etwad anders...
-{}
-"""
-
-RESTO_TABLES = {
-    "soup_table": {"soup"},
-    "meat_table": {"meat"},
-    "fish_table": {"fish"},
-    "vegi_table": {"vegetarian", "vegan"},
-}
-
-
-@app.route("/resto", methods=["GET"])
-def resto_menu():
-    today = datetime.today()
-    url = f"https://hydra.ugent.be/api/2.0/resto/menu/nl/{today.year}/{today.month}/{today.day}.json"
-    resto = requests.get(url, timeout=5).json()
-
-    if not resto["open"]:
-        return "De resto is vandaag gesloten."
-    else:
-
-        def table_for(kinds):
-            items = [meal for meal in resto["meals"] if meal["kind"] in kinds]
-            return format_items(items)
-
-        def format_items(items):
-            if not items:
-                return "None :("
-            maxwidth = max(len(item["name"]) for item in items)
-            return "\n".join(
-                "{name: <{width}}{price}".format(
-                    name=item["name"], width=maxwidth + 2, price=item["price"]
-                )
-                for item in items
-            )
-
-        recognized_kinds = set.union(*RESTO_TABLES.values())
-        uncategorized_meals = [
-            meal for meal in resto["meals"] if meal["kind"] not in recognized_kinds
-        ]
-
-        uncategorized = (
-            ""
-            if not uncategorized_meals
-            else UNCATEGORIZED_TEMPLATE.format(
-                format_items(
-                    [
-                        # Small hack: change "name" into "name (kind)"
-                        {**meal, "name": "{} ({})".format(meal["name"], meal["kind"])}
-                        for meal in uncategorized_meals
-                    ]
-                )
-            )
-        )
-
-        return RESTO_TEMPLATE.format(
-            **{k: table_for(v) for k, v in RESTO_TABLES.items()},
-            uncategorized=uncategorized,
-            vegetable_table="\n".join(resto["vegetables"]),
-        )
-
-
-@app.route("/resto.json", methods=["GET"])
-def resto_menu_json():
-    return mattermost_response(resto_menu(), ephemeral=True)
