+++ app/routes/__init__.py
@@ -1,0 +1,8 @@
+from cammie import cammie_blueprint
+from door_access import door_access_blueprint
+from door_control import door_control_blueprint
+from doorkeeper import doorkeeper_blueprint
+from fingerprint import fingerprint_blueprint
+from quote import quote_blueprint
+from resto import resto_blueprint
+from spaceapi import spaceapi_blueprint
